pyproject.toml
[project]
name = "prompt-engineering-research"
version = "1.0.0"
description = "Graduate-level research project on prompt engineering techniques"
requires-python = ">=3.10"
authors = [
    {name = "Graduate Research Team"}
]
readme = "README.md"

dependencies = [
    # LLM Providers
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "google-generativeai>=0.3.0",
    "ollama>=0.1.6",
    "requests>=2.31.0",

    # ML & Embeddings
    "sentence-transformers>=2.5.0",
    "scikit-learn>=1.4.0",
    "torch>=2.0.0",
    "transformers>=4.36.0",

    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",

    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "tqdm>=4.66.0",
    "pydantic>=2.5.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
